+++ src/llvm-remove-addrspaces.cpp
@@ -14,6 +14,7 @@
 
 #include "codegen_shared.h"
 #include "julia.h"
+#include "passes.h"
 
 #define DEBUG_TYPE "remove_addrspaces"
 
@@ -231,18 +232,7 @@ unsigned removeAllAddrspaces(unsigned AS)
     return AddressSpace::Generic;
 }
 
-struct RemoveAddrspacesPass : public ModulePass {
-    static char ID;
-    AddrspaceRemapFunction ASRemapper;
-    RemoveAddrspacesPass(
-            AddrspaceRemapFunction ASRemapper = removeAllAddrspaces)
-        : ModulePass(ID), ASRemapper(ASRemapper){};
-
-public:
-    bool runOnModule(Module &M) override;
-};
-
-bool RemoveAddrspacesPass::runOnModule(Module &M)
+bool removeAddrspaces(Module &M, AddrspaceRemapFunction ASRemapper)
 {
     ValueToValueMapTy VMap;
     AddrspaceRemoveTypeRemapper TypeRemapper(ASRemapper);
@@ -457,8 +447,22 @@ bool RemoveAddrspacesPass::runOnModule(Module &M)
     return true;
 }
 
-char RemoveAddrspacesPass::ID = 0;
-static RegisterPass<RemoveAddrspacesPass>
+
+struct RemoveAddrspacesPassLegacy : public ModulePass {
+    static char ID;
+    AddrspaceRemapFunction ASRemapper;
+    RemoveAddrspacesPassLegacy(
+            AddrspaceRemapFunction ASRemapper = removeAllAddrspaces)
+        : ModulePass(ID), ASRemapper(ASRemapper){};
+
+public:
+    bool runOnModule(Module &M) override {
+        return removeAddrspaces(M, ASRemapper);
+    }
+};
+
+char RemoveAddrspacesPassLegacy::ID = 0;
+static RegisterPass<RemoveAddrspacesPassLegacy>
         X("RemoveAddrspaces",
           "Remove IR address space information.",
           false,
@@ -467,7 +471,17 @@ static RegisterPass<RemoveAddrspacesPass>
 Pass *createRemoveAddrspacesPass(
         AddrspaceRemapFunction ASRemapper = removeAllAddrspaces)
 {
-    return new RemoveAddrspacesPass(ASRemapper);
+    return new RemoveAddrspacesPassLegacy(ASRemapper);
+}
+
+RemoveAddrspacesPass::RemoveAddrspacesPass() : RemoveAddrspacesPass(removeAllAddrspaces) {}
+
+PreservedAnalyses RemoveAddrspacesPass::run(Module &M, ModuleAnalysisManager &AM) {
+    if (removeAddrspaces(M, ASRemapper)) {
+        return PreservedAnalyses::allInSet<CFGAnalyses>();
+    } else {
+        return PreservedAnalyses::all();
+    }
 }
 
 
@@ -483,16 +497,16 @@ unsigned removeJuliaAddrspaces(unsigned AS)
         return AS;
 }
 
-struct RemoveJuliaAddrspacesPass : public ModulePass {
+struct RemoveJuliaAddrspacesPassLegacy : public ModulePass {
     static char ID;
-    RemoveAddrspacesPass Pass;
-    RemoveJuliaAddrspacesPass() : ModulePass(ID), Pass(removeJuliaAddrspaces){};
+    RemoveAddrspacesPassLegacy Pass;
+    RemoveJuliaAddrspacesPassLegacy() : ModulePass(ID), Pass(removeJuliaAddrspaces){};
 
-    bool runOnModule(Module &M) { return Pass.runOnModule(M); }
+    bool runOnModule(Module &M) override { return Pass.runOnModule(M); }
 };
 
-char RemoveJuliaAddrspacesPass::ID = 0;
-static RegisterPass<RemoveJuliaAddrspacesPass>
+char RemoveJuliaAddrspacesPassLegacy::ID = 0;
+static RegisterPass<RemoveJuliaAddrspacesPassLegacy>
         Y("RemoveJuliaAddrspaces",
           "Remove IR address space information.",
           false,
@@ -500,7 +514,11 @@ static RegisterPass<RemoveJuliaAddrspacesPass>
 
 Pass *createRemoveJuliaAddrspacesPass()
 {
-    return new RemoveJuliaAddrspacesPass();
+    return new RemoveJuliaAddrspacesPassLegacy();
+}
+
+PreservedAnalyses RemoveJuliaAddrspacesPass::run(Module &M, ModuleAnalysisManager &AM) {
+    return RemoveAddrspacesPass(removeJuliaAddrspaces).run(M, AM);
 }
 
 extern "C" JL_DLLEXPORT void LLVMExtraAddRemoveJuliaAddrspacesPass_impl(LLVMPassManagerRef PM)